style: rename & format code
